build: consolidate TypeScript config and ensure that type defintions are being included in dist
The previous change to include type definitions in dist (7c4a296) was incomplete.
This change adds some further refinements to ensure that a .d.ts file is included in dist, allowing consumers to work with typesafe code from the pundit package in TypeScript projects.
I have reverted our TypeScript configuration back to a single tsconfig.json for simplicity and to avoid duplication, while setting up a separate config for the examples to keep those separate.

This change also upgrades the project to TypeScript 6 and adds some neccessary changes to the config.
Note: we need set compilerOptions->strict to false on this project due to the approach of adding the policy actions being dynamically added to the policy class at runtime.
