+++ examples/extended-policy-example.tsx
@@ -4,9 +4,8 @@ import { AuthorisableUser, AuthorisablePost } from './types'
 import { author, admin, post } from './example-data'
 
 class PostPolicy extends Policy {
-  user: AuthorisableUser
-
-  record: AuthorisablePost
+  declare user: AuthorisableUser
+  declare record: AuthorisablePost
 
   constructor(user: AuthorisableUser, record: AuthorisablePost) {
     super(user, record)
