+++ examples/package.json
@@ -5,7 +5,8 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vite build"
+    "build": "tsc && vite build",
+    "types": "tsc"
   },
   "dependencies": {
     "react": "^19.2.4",
