+++ src/react/when.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
 import { expect } from 'vitest'
-import Policy from '../../src/policy'
-import When from '../../src/react/when'
+import Policy from '../policy'
+import When from './when'
 
 describe('<When />', () => {
   describe('policy parameter', () => {
