--- tsconfig.build.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "./dist",
-    "baseUrl": "./src",
-    "jsx": "react",
-    "esModuleInterop": true,
-    "strictNullChecks": true,
-    "skipLibCheck": true,
-    "types": ["vitest/globals"]
-  },
-  "include": ["src/**/*"]
-}
